docs: clarify unified starter question util
